Final data model changes (for real)
Section has a list of list of links, which run parallel to paragraphs. This gives us a relative ordering of the links and allows us to see which links are in a paragraph. 

Links contain page ids as well, this allows us to reduce database queries when going from link directly to page.

Aliases have a list of links, this allows us to build the wiki link table significantly faster.