Add notes to data model
Notes reside in a section and contain a default empty string value. 

When paragraphs are added, an empty note is also added. When a paragraph is deleted, the corresponding note is also removed.

They are considered "used" when they have actual content. When retrieving a section, notes are only included if they are used.